COM-1068 change seed
