+++ tests/seed/companySeed.js
@@ -1,0 +1,20 @@
+const { ObjectID } = require('mongodb');
+
+const authCompany = {
+  _id: new ObjectID(),
+  name: 'Test SAS',
+  tradeName: 'Test',
+  prefixNumber: 101,
+  iban: '1234',
+  bic: '5678',
+  ics: '9876',
+  folderId: '0987654321',
+  directDebitsFolderId: '1234567890',
+  customersFolderId: 'mnbvcxz',
+  auxiliariesFolderId: 'iuytre',
+  customersConfig: {
+    billingPeriod: 'two_weeks',
+  },
+};
+
+module.exports = { authCompany };
